WIP - interp: Allocate memory for constants.
The code is generated with the expectation of using addresses of values/constants for arrays, closures, etc..
So, the previous hack for closures to store the label cannot work in general.

The code has been changed to allocate a memory address for each constant and use these addresses as the codegen expects.
There is indirection on float constants, function addresses in closures and so on. This is now modelled in the interpreter.